fix: use an actor check that exists on Flarum 2.0.0-rc.8
Updating the extension took the whole forum down with a 500:

  BadMethodCallException: Call to undefined method Flarum\User\Guest::isRegistered()
  .../flarum-extension/src/Api/UserResourceFields.php:30

The isRegistered() helper does not exist in Flarum 2.0.0-rc.8 - it is a later 2.x
addition - and the local framework checkout used while writing this code came from
the 2.x branch, which is newer than the deployment. The call also sat inside a
visible() callback, so every request that serialized a user failed: the index
page, the discussion list, everything. The log shows it failing for
Flarum\User\Guest and Flarum\User\User alike.

The portable test is the actor id. In rc.8 a guest is a Flarum\User\Guest with
`public int $id = 0`, and because Guest *extends* User an instanceof check is not
a substitute, while a real account always has a positive id. Both occurrences now
read

  (int) ($context->getActor()?->id ?? 0) > 0

Hardened while in there: Carbon 3 returns a float from diffInSeconds and intdiv()
takes ints, so the status page casts once before dividing instead of relying on a
deprecated implicit conversion.

Process fix: tools/verify.mjs section 19 bans isRegistered() outright and asserts
that composer.json's flarum/core constraint covers the release the code is audited
against, with that release written into the check. The lesson is recorded there:
auditing against a local framework checkout that is newer than the deployment is
not verification.

Bumped to 0.0.4 and released, so restoring the forum is a composer update.

+++ flarum-extension/src/Api/Controller/StatusPageController.php
@@ -219,7 +219,9 @@ HTML;
             return $e($t('never'));
         }
 
-        $seconds = max(0, $moment->diffInSeconds(Carbon::now()));
+        // Carbon 3 returns a float from diffInSeconds and intdiv() takes ints, so
+        // the value is cast once here rather than at three call sites.
+        $seconds = (int) max(0, $moment->diffInSeconds(Carbon::now()));
 
         if ($seconds < 60) {
             return $e(sprintf($t('seconds_ago'), $seconds));
